requirements.txt
typing>=3.7.4.3
python-dotenv>=1.0.1
ping3>=4.0.8
pysmb>=1.2.9.1
ldap3>=2.9.1
asyncio>=3.4.3
aiofiles>=24.1.0
smbprotocol>=1.14.0
colorlog>=6.8.2
chardet>=5.2.0
aiomysql>=0.2.0